// examples/with-design-system/libs/cognito-auth/src/lib/Auth.interface.tsx
import { TExtendedUser } from '@flounder/contracts';
import { CognitoApi } from './CognitoApi/CognitoApi';

export enum CognitoErrorCode {
  NotAuthorizedException = 'NotAuthorizedException',
  UserNotConfirmedException = 'UserNotConfirmedException',
  CodeMismatchException = 'CodeMismatchException',
  ExpiredCodeException = 'ExpiredCodeException',
  InvalidPasswordException = 'InvalidPasswordException',
  InvalidParameterException = 'InvalidParameterException',
  LimitExceededException = 'LimitExceededException',
}

export enum CognitoContext {
  ResetPassword = 'ResetPassword',
}

export enum CognitoAppErrorMessages {
  UserNotConfirmed = 'User account not confirmed. Check your mail',
  General = 'Something goes wrong',
  InvalidUserOrPassword = 'Invalid credentials',
  CodeMismatchException = 'Wrong verification code',
  ExpiredCodeException = 'Code expired. Resend new code',
}

export interface IUserNotConfirmedError extends CognitoError {
  email: string;
}

export class UserNotConfirmedError implements IUserNotConfirmedError {
  constructor(
    public code: CognitoErrorCode,
    public message: string,
    public email: string
  ) {}
}

export interface CognitoError {
  code: CognitoErrorCode;
  message: string;
}

export interface AuthProviderProps {
  cognitoApi: CognitoApi;
  ssrUser: TExtendedUser | null;
}

export type IAuthContext = {
  cognitoApi?: CognitoApi;
  currentUser: TExtendedUser | null;
  setCurrentAuthenticatedUser: () => Promise<void>;
};